+++ packages/reporter/src/main.tsx
@@ -47,12 +47,12 @@ export interface BaseReporterProps {
   runnerStore: MobxRunnerStore
 }
 
-export interface SingleReporterProps extends BaseReporterProps{
+export interface SingleReporterProps extends BaseReporterProps {
   runMode?: 'single'
 }
 
 // In React Class components (now deprecated), we used to use appState as a default prop. Now since defaultProps are not supported in functional components, we can use ES6 default params to accomplish the same thing
-const Reporter: React.FC<SingleReporterProps> = observer(({ appState = appStateDefault, runner, className, error, runMode = 'single', studioEnabled, autoScrollingEnabled, isSpecsListOpen, resetStatsOnSpecChange, renderReporterHeader = (props: ReporterHeaderProps) => <Header {...props}/>, runnerStore }) => {
+const Reporter: React.FC<SingleReporterProps> = observer(({ appState = appStateDefault, runner, className, error, runMode = 'single', studioEnabled, autoScrollingEnabled, isSpecsListOpen, resetStatsOnSpecChange, renderReporterHeader = (props: ReporterHeaderProps) => <Header {...props} />, runnerStore }) => {
   const previousSpecRunId = usePrevious(runnerStore.specRunId)
   const [isMounted, setIsMounted] = useState(false)
   const [isInitialized, setIsInitialized] = useState(false)
@@ -101,7 +101,7 @@ const Reporter: React.FC<SingleReporterProps> = observer(({ appState = appStateD
     runnablesStore.setRunningSpec(runnerStore.spec.relative)
     if (
       resetStatsOnSpecChange &&
-        runnerStore.specRunId !== previousSpecRunId
+      runnerStore.specRunId !== previousSpecRunId
     ) {
       statsStore.reset()
     }
@@ -112,7 +112,7 @@ const Reporter: React.FC<SingleReporterProps> = observer(({ appState = appStateD
       'studio-active': appState.studioActive,
       'mounted': isMounted,
     })}>
-      {renderReporterHeader({ appState, statsStore, runnablesStore })}
+      {renderReporterHeader({ appState, statsStore, runnablesStore, spec: runnerStore.spec })}
       {appState?.isPreferencesMenuOpen ? (
         <TestingPreferences appState={appState} />
       ) : (